+++ src/compas_timber/elements/wall.py
@@ -7,14 +7,43 @@ from compas.geometry import Polyline
 from compas.geometry import bounding_box
 
 from .slab import Slab
-from compas_timber.utils import classify_polyline_segments
+
 
 class OpeningType(object):
+    """Constants for different types of openings in walls.
+
+    Attributes
+    ----------
+    DOOR : str
+        Constant for door openings.
+    WINDOW : str
+        Constant for window openings.
+    """
+
     DOOR = "door"
     WINDOW = "window"
 
 
 class Opening(Data):
+    """Represents an opening in a wall (door, window, etc.).
+
+    Parameters
+    ----------
+    polyline : :class:`~compas.geometry.Polyline`
+        The polyline defining the boundary of the opening.
+    opening_type : str
+        The type of opening (e.g., "door", "window").
+    **kwargs : dict, optional
+        Additional keyword arguments.
+
+    Attributes
+    ----------
+    polyline : :class:`~compas.geometry.Polyline`
+        The polyline defining the boundary of the opening.
+    opening_type : str
+        The type of opening.
+    """
+
     @property
     def __data__(self):
         return {
@@ -31,17 +60,72 @@ class Opening(Data):
         return "Opening(type={})".format(self.opening_type)
 
     def orient_polyline(self, normal):
+        """Orient the polyline consistently with the given normal vector.
+
+        Parameters
+        ----------
+        normal : :class:`~compas.geometry.Vector`
+            The normal vector to orient the polyline with.
+        """
         self.polyline = _oriented_polyline(self.polyline, normal)
 
 
 class Wall(Slab):
     """Represents a single timber wall element.
-    Serves as container for beams joints and other related elements and groups them together to form a wall.
 
+    Serves as container for beams, joints and other related elements and groups them together to form a wall.
     Wall is often a single unit of prefabricated timber wall element.
-    It is often refered to as an enveloping body.
+    It is often referred to as an enveloping body.
 
-    TODO: complete this docstring
+    Parameters
+    ----------
+    frame : :class:`~compas.geometry.Frame`
+        The coordinate system (frame) of this wall.
+    length : float
+        Length of the wall.
+    width : float
+        Width of the wall.
+    thickness : float
+        Thickness of the wall.
+    outline_a : :class:`~compas.geometry.Polyline`, optional
+        A polyline representing the principal outline of this wall.
+    outline_b : :class:`~compas.geometry.Polyline`, optional
+        A polyline representing the associated outline of this wall.
+    openings : list[:class:`~compas_timber.elements.Opening`], optional
+        A list of Opening objects representing openings in this wall.
+    name : str, optional
+        Name of the wall. Defaults to "Wall".
+    **kwargs : dict, optional
+        Additional keyword arguments.
+
+    Attributes
+    ----------
+    frame : :class:`~compas.geometry.Frame`
+        The coordinate system (frame) of this wall.
+    outline : :class:`~compas.geometry.Polyline`
+        The outline of the wall (same as outline_a).
+    thickness : float
+        Thickness of the wall.
+    name : str
+        Name of the wall.
+    attributes : dict
+        Dictionary of additional attributes.
+    is_wall : bool
+        Always True for walls.
+    origin : :class:`~compas.geometry.Point`
+        The origin point of the wall frame.
+    centerline : :class:`~compas.geometry.Line`
+        The centerline of the wall (alias for baseline).
+    baseline : :class:`~compas.geometry.Line`
+        The baseline of the wall from first to second point of outline.
+    corners : tuple[:class:`~compas.geometry.Point`, ...]
+        The 8 corner points of the wall bounding box.
+    faces : tuple[:class:`~compas.geometry.Frame`, ...]
+        The 6 faces of the wall as frames (bottom, left, top, right, back, front).
+    end_faces : tuple[:class:`~compas.geometry.Frame`, :class:`~compas.geometry.Frame`]
+        The back and front faces of the wall.
+    envelope_faces : tuple[:class:`~compas.geometry.Frame`, ...]
+        The envelope faces of the wall (bottom, left, top, right).
 
     """
 
@@ -55,7 +139,7 @@ class Wall(Slab):
         return data
 
     def __init__(self, frame, length, width, thickness, outline_a=None, outline_b=None, openings=None, name=None, **kwargs):
-        print("args",frame, length, width, thickness)
+        print("args", frame, length, width, thickness)
         super(Wall, self).__init__(frame, length, width, thickness, outline_a, outline_b, openings, **kwargs)
         self.outline = outline_a
         self.attributes = {}
@@ -72,26 +156,63 @@ class Wall(Slab):
 
     @property
     def is_wall(self):
+        """Check if this element is a wall.
+
+        Returns
+        -------
+        bool
+            Always True for walls.
+        """
         return True
 
     @property
     def origin(self):
+        """The origin point of the wall frame.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Point`
+            The origin point of the wall frame.
+        """
         assert self.frame
         return self.frame.point.copy()
 
     @property
     def centerline(self):
+        """The centerline of the wall.
+
+        This is an alias for baseline to maintain compatibility with topology finding algorithms.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Line`
+            The centerline of the wall.
+        """
         # TODO: temp hack to make this compatible with `find_topology`.
         return self.baseline
 
     @property
     def baseline(self):
+        """The baseline of the wall from first to second point of outline.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Line`
+            The baseline of the wall.
+        """
         # type: () -> Line
         points = self.outline.points
         return Line(points[0], points[1])
 
     @property
     def corners(self):
+        """The 8 corner points of the wall bounding box.
+
+        Returns
+        -------
+        tuple[:class:`~compas.geometry.Point`, ...]
+            The 8 corner points in order consistent with Box.from_bounding_box requirements.
+        """
         assert self.frame
         if not self._corners:
             points = self.outline.points
@@ -109,6 +230,13 @@ class Wall(Slab):
 
     @property
     def faces(self):
+        """The 6 faces of the wall as frames.
+
+        Returns
+        -------
+        tuple[:class:`~compas.geometry.Frame`, ...]
+            The 6 faces as frames in BTLx ref-side system order: (bottom, left, top, right, back, front).
+        """
         if not self._faces:
             corners = self.corners
             bottom_face = Frame.from_points(corners[0], corners[1], corners[3])
@@ -123,23 +251,73 @@ class Wall(Slab):
 
     @property
     def end_faces(self):
+        """The back and front faces of the wall.
+
+        Returns
+        -------
+        tuple[:class:`~compas.geometry.Frame`, :class:`~compas.geometry.Frame`]
+            The back and front faces of the wall.
+        """
         return self.faces[-2:]
 
     @property
     def envelope_faces(self):
+        """The envelope faces of the wall (bottom, left, top, right).
+
+        Returns
+        -------
+        tuple[:class:`~compas.geometry.Frame`, ...]
+            The envelope faces of the wall.
+        """
         return self.faces[:4]
 
     def compute_geometry(self, _=False):
+        """Compute the geometry of the wall.
+
+        Parameters
+        ----------
+        _ : bool, optional
+            Unused parameter for compatibility. Defaults to False.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Brep`
+            The geometry of the wall as a Brep.
+        """
         assert self.frame
 
         extrusion_vector = self.frame.zaxis * self.thickness
         return Brep.from_extrusion(self.outline, extrusion_vector)
 
     def compute_aabb(self, inflate_by=0.1):
+        """Compute the Axis Aligned Bounding Box (AABB) of the wall.
+
+        Parameters
+        ----------
+        inflate_by : float, optional
+            Amount to inflate the bounding box by. Defaults to 0.1.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Box`
+            The AABB of the wall.
+        """
         obb = self.compute_obb(inflate_by)
         return Box.from_bounding_box(bounding_box(obb.points))
 
     def compute_obb(self, inflate_by=0.0):
+        """Compute the Oriented Bounding Box (OBB) of the wall.
+
+        Parameters
+        ----------
+        inflate_by : float, optional
+            Amount to inflate the bounding box by. Defaults to 0.0.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Box`
+            The OBB of the wall.
+        """
         assert self.frame
         # TODO: this is more like obb than aabb
         box = Box.from_bounding_box(self.corners)
@@ -149,65 +327,65 @@ class Wall(Slab):
         return box
 
     def rotate(self):
+        """Rotate the wall by shifting the outline points and updating the frame.
+
+        This method rotates the wall outline by one position and updates the frame accordingly.
+        """
         assert self.frame
         self.outline = Polyline(self.outline.points[1:] + [self.outline.points[1]])
         assert self.outline.is_closed
-        self.frame = Slab._frame_from_polyline(self.outline, self.frame.normal)
+        self.frame = self._frame_from_polyline(self.outline, self.frame.normal)
         assert len(self.outline) == 5
-
-    def __repr__(self):
-        return "Slab(name={}, {}, {}, {:.3f})".format(self.name, self.frame, self.outline, self.thickness)
 
     @staticmethod
     def _frame_from_polyline(polyline, normal):
+        """Create a frame from a polyline and normal vector.
+
+        Parameters
+        ----------
+        polyline : :class:`~compas.geometry.Polyline`
+            The polyline to create the frame from.
+        normal : :class:`~compas.geometry.Vector`
+            The normal vector for the frame.
+
+        Returns
+        -------
+        :class:`~compas.geometry.Frame`
+            A frame with origin at the first point of the polyline.
+        """
         points = polyline.points
         xaxis = points[1] - points[0]
         xaxis.unitize()
         yaxis = normal.cross(xaxis)
         return Frame(points[0], xaxis, yaxis)
 
-    # @classmethod
-    # def from_boundary(cls, polyline, normal, thickness, openings=None, **kwargs):
-    #     """Use this to make sure the polyline is oriented correctly."""
-    #     oriented_polyline = _oriented_polyline(polyline, normal)
-    #     openings = openings or []
-    #     for opening in openings:
-    #         opening.orient_polyline(normal)
-    #     wall_frame = cls._frame_from_polyline(oriented_polyline, normal)
-    #     return cls(oriented_polyline, thickness, openings, wall_frame, **kwargs)
-
-    # @classmethod
-    # def from_brep(cls, brep, thickness, **kwargs):
-    #     """Creates a wall from a brep with a single planar face."""
-    #     if len(brep.faces) > 1:
-    #         raise ValueError("Can only single-face breps to create a Wall. This brep has {}".format(len(brep.faces)))
-
-    #     # trims are oriented consistently, depending on the face orientation
-    #     face = brep.faces[0]
-    #     trims = face.boundary.trims
-    #     winding_direction = "cw" if face.is_reversed else "ccw"
-
-    #     # separate the outline from the cutouts (concave)
-    #     # these are still part of the boundary loop but we cound them as openings (doors)
-    #     boundary = Polyline([t.start_vertex.point for t in trims] + [trims[-1].end_vertex.point])
-    #     face_frame = face.frame_at(0, 0)
-    #     outline_vertices, internal_groups = classify_polyline_segments(boundary, normal=face_frame.normal, direction=winding_direction)
-    #     outline = Polyline([boundary[i] for i in outline_vertices])
-
-    #     openings = []
-    #     for group in internal_groups:
-    #         points = [boundary[i] for i in group]
-    #         openings.append(Opening(Polyline(points), OpeningType.DOOR))
-
-    #     # internal cuts (windows) are not part of the outline and can be fetched from the loops that are not boundary
-    #     for hole in face.holes:
-    #         points = [t.start_vertex.point for t in hole.trims] + [hole.trims[-1].end_vertex.point]
-    #         openings.append(Opening(Polyline(points), OpeningType.WINDOW))
-
-    #     return cls.from_boundary(outline, face_frame.normal, thickness, openings, **kwargs)
-
 
 def _oriented_polyline(polyline, normal):
+    """Return a polyline that is oriented consistently counterclockwise around the normal.
+
+    Parameters
+    ----------
+    polyline : :class:`~compas.geometry.Polyline`
+        The input polyline to orient.
+    normal : :class:`~compas.geometry.Vector`
+        The normal vector to orient around.
+
+    Returns
+    -------
+    :class:`~compas.geometry.Polyline`
+        A polyline oriented counterclockwise around the normal.
+
+    Notes
+    -----
+    The function assumes a specific orientation:
+
+    .. code-block:: text
+
+        ^  3 ---- 2
+        |  |      |
+        z  0 ---- 1
+           x -->
+    """
     # returns a polyline that is oriented consistently ccw around the normal
     # ^  3 ---- 2
     # |  |      |